+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Support table editing via `date_time` for date/time values
 - Support table default sort configuration via sort_property
+- Support table default sort configuration via sort block
+- Support table default sort configuration via sort_by block
 
 ### 4.17.9.0